test(memory-core): relocate concurrency test to canonical path and refactor to use production lock logic (#10693)
Also added CHECK constraint to SummarizationJobs status column.
